changed amount field to BigDecimal

+++ src/main/java/com/talentwunder/financetracker/model/request/TransactionRequest.java
@@ -2,8 +2,10 @@ package com.talentwunder.financetracker.model.request;
 
 import com.talentwunder.financetracker.model.entity.TransactionType;
 
+import java.math.BigDecimal;
+
 public interface TransactionRequest {
     TransactionType getType();
-    Double getAmount();
+    BigDecimal getAmount();
     String getDescription();
 }
